+++ src/xilinx.cpp
@@ -217,7 +217,7 @@ void Xilinx::program_mem(BitParser &bitfile)
  *        so to send only a cmd set len to 0 (or omit this param)
  */
 int Xilinx::spi_put(uint8_t cmd,
-			uint8_t *tx, uint8_t *rx, uint16_t len)
+			uint8_t *tx, uint8_t *rx, uint32_t len)
 {
 	int xfer_len = len + 1 + ((rx == NULL) ? 0 : 1);
 	uint8_t jtx[xfer_len];
@@ -243,7 +243,7 @@ int Xilinx::spi_put(uint8_t cmd,
 	return 0;
 }
 
-int Xilinx::spi_put(uint8_t *tx, uint8_t *rx, uint16_t len)
+int Xilinx::spi_put(uint8_t *tx, uint8_t *rx, uint32_t len)
 {
 	int xfer_len = len + ((rx == NULL) ? 0 : 1);
 	uint8_t jtx[xfer_len];
